reorg legacy detection code
